feat: implement log levels
