+++ .env
@@ -13,5 +13,5 @@ JWT_SECRET_KEY=secret123123
 JWT_SECRET_REFRESH_KEY=secret123123
 TOKEN_EXPIRE_TIME=1h
 TOKEN_REFRESH_EXPIRE_TIME=24h
-LOG_LEVEL=2
-LOG_SIZE_KB=5
+LOG_LEVEL=1
+LOG_SIZE_KB=50
